% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{bbox_to_SpatialPolygons}
\alias{bbox_to_SpatialPolygons}
\title{Convert a 4-element bbox vector to a SpatialPolygons object}
\usage{
bbox_to_SpatialPolygons(bb = c(-72, -63, 39, 46),
  proj_string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
}
\arguments{
\item{bb}{a 4-element numeric vector of left, bottom, right, top coordinates}

\item{proj_string}{a proj4string suitable to pass to \code{sp::CRS()}}
}
\value{
a SpatialPolygons object
}
\description{
Convert a 4-element bbox vector to a SpatialPolygons object
}

