% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{bbox_to_SpatialPolygonsDataFrame}
\alias{bbox_to_SpatialPolygonsDataFrame}
\title{Convert a 4-element bbox vector to a SpatialPolygonsDataFrame object}
\usage{
bbox_to_SpatialPolygonsDataFrame(bb = c(-72, -63, 39, 46), ...)
}
\arguments{
\item{bb}{a 4-element numeric vector of left, bottom, right, top coordinates}

\item{...}{further arguments for \code{bbox_to_SpatialPolygons}}
}
\value{
a SpatialPolygons object
}
\description{
Convert a 4-element bbox vector to a SpatialPolygonsDataFrame object
}

