% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{find_nearth_rasters}
\alias{find_nearth_rasters}
\title{Find the full path for a raster dataset}
\usage{
find_nearth_rasters(name = "GRAY_50M_SR_O", ext = ".tif",
  path = nearth_path(what = "raster"))
}
\arguments{
\item{name}{character, one or more names to find}

\item{ext}{character the file extension to seek, by default '.tif'}

\item{path}{character the path to the Natural Earth raster datasets}
}
\value{
character, named logical, TRUE if the file exists
}
\description{
Find the full path for a raster dataset
}

