% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{find_nearth_vectors}
\alias{find_nearth_vectors}
\title{Find the full path for a vector dataset}
\usage{
find_nearth_vectors(name = "ne_50m_coastline", ext = ".shp",
  path = nearth_path(what = "vector"))
}
\arguments{
\item{name}{character, one or more names to find}

\item{ext}{character the file extension to seek, by default '.shp'}

\item{path}{character the path to the Natural Earth vector datasets}
}
\value{
a named character vector, missing files are returned as empty character
}
\description{
Find the full path for a vector dataset
}

