% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nearth.R
\name{has_nearth}
\alias{has_nearth}
\title{Test for the existence of a Natural Earth path in R's \code{options()}}
\usage{
has_nearth(what = c("vector", "raster")[1])
}
\arguments{
\item{what}{character, either 'vector' (default) or 'raster'}
}
\value{
named logical, TRUE of the path exists
}
\description{
Test for the existence of a Natural Earth path in R's \code{options()}
}

