% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth-package.R
\docType{package}
\name{nearth-package}
\alias{nearth}
\alias{nearth-package}
\title{nearth: Natural Earth local data access}
\description{
Natural Earth \url{http://www.naturalearthdata.com/} provides beautiful 
 physical and cultural data in both vector and raster formats.  This package 
 local storage.
}

