% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nearth.R
\name{nearth_path}
\alias{nearth_path}
\title{Retrieve the Natural Earth path}
\usage{
nearth_path(what = c("vector", "raster")[1])
}
\arguments{
\item{what}{character, either 'vector' (default) or 'raster'}
}
\value{
the Natural Earth path or ""
}
\description{
Retrieve the Natural Earth path
}

