% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nearth.R
\name{read_nearth}
\alias{read_nearth}
\title{Read one or more Natural Earth data files.}
\usage{
read_nearth(name = "ne_50m_coastline", what = c("vector", "raster")[1], ...)
}
\arguments{
\item{name}{character, one or more names to find, must point to either
vectors or rasters but not a mix of the two}

\item{what}{character either 'vector' (default) or 'raster'}

\item{...}{further arguments for \code{nearth::read_nearth_raster()} or
or \code{nearth::read_nearth_vector()}}
}
\description{
Read one or more Natural Earth data files.
}

