% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{read_nearth_vector}
\alias{read_nearth_vector}
\title{Read one vector data set using \code{rgdal::readOGR()}}
\usage{
read_nearth_vector(filename, bb = NULL, ...)
}
\arguments{
\item{filename}{character, one file name}

\item{bb}{bounding box to clip to [left, right, bottom, top]}

\item{...}{further arguments for \code{rgdal::readOGR()}}
}
\value{
Spatial* class object or NULL
}
\description{
Read one vector data set using \code{rgdal::readOGR()}
}

