% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nearth.R
\name{strip_extension}
\alias{strip_extension}
\title{Strip the extension (e.g. ".ext") off one or more filenames}
\usage{
strip_extension(x, sep = ".")
}
\arguments{
\item{x}{character one or more filenames}

\item{sep}{character, separator between name of extension}
}
\value{
named character vector
}
\description{
Strip the extension (e.g. ".ext") off one or more filenames
}

