% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nearth.R
\name{read_nearth_vector}
\alias{read_nearth_vector}
\title{Read one vector data set using \code{rgdal::readOGR()}}
\usage{
read_nearth_vector(filename, ...)
}
\arguments{
\item{filename}{character, one file name}

\item{...}{further arguments for \code{rgdal::readOGR()}}
}
\value{
Spatial* class object or NULL
}
\description{
Read one vector data set using \code{rgdal::readOGR()}
}

